must not have ...."landing" page.. in the background; that will get called *before* the users are defined.
We must have the users defined so before any menus are rendered, else the 'login' menu won't know if anyone is logged in.  So, because of the way cucumber sets things up and executes things, the statement 'And I am on the "landing" page' must be down in the Scenario, and not one of the Background steps.