use the method from User; User should be responsible for knowing how to check, not a view
Let User be responsible for doing whatever is needed to figure out if they are are member. Do not directly check the 'internals' (an attribute) of the User class.  A User may change how this works (exactly what is invovled in determining if a User is a member).  No other class or object should have to know exactly what is invovled -- which attributes are/aren't invovled, etc.  All they should care about is the answer to the question: 'are you a member?'