Redirects after login (#75)
* added a route for the info page, routes after_sign_in path there for users who are not admins, conditionally displays the content of that page

* simplified the controller method

* simplified the admin chek in user model

* added acceptance tests for logging in as admin

* didnt need self

* fixed stuff on index after comments form SHF

* added text for the information page, drye'd it up a bit

* scenarios to make sure the right page is shown after logging in

* happy and sad path runs green

* add steps to be on and to check the 'member instructions' and 'user instructions' page

* also test that we are on the correct page (the user instructions/member instructions)

* use the method from User; User should be responsible for knowing how to check, not a view

Let User be responsible for doing whatever is needed to figure out if they are are member. Do not directly check the 'internals' (an attribute) of the User class.  A User may change how this works (exactly what is invovled in determining if a User is a member).  No other class or object should have to know exactly what is invovled -- which attributes are/aren't invovled, etc.  All they should care about is the answer to the question: 'are you a member?'