BUG: companies with members query wrong (#870)
* scope :current_membership only looks for accepted application and current payments

Payments can only be made after all membership requirements are satisified.  If a member must agree to the Ethical Guidelines, they they must do that before a payment can be made.  Thus if a user has current payments (term expiriation >= today), then this implies the member has agreed to Ethical Guidelines if they are required to.

* fix feature: membership gets email only for new members

* (minor) clarify scenario descriptions

* show payments to admin on Payor (Users, Companies) pages